requirements.txt
Django==5.2.8
pydantic==2.12.4
pydantic-settings==2.12.0
pytest==9.0.1
pytest-django==4.11.1
psycopg2-binary==2.9.11
django-ninja==1.5.0
